requirements.txt
openai
streamlit
pandas
python-dotenv